LayoutTests/inspector/extensions.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test2.js"></script>
<script src="extensions-test.js"></script>
<script type="text/javascript">

window.inspectedValue = { str: "foo", num: 42 };

function extension_testGetAllResources(nextTest)
{
    function compareResources(a, b)
    {
        return a.har.request.url.localeCompare(b.har.request.url);
    }

    function onResource(result)
    {
        var resources = result.sort(compareResources);

        for (var i = 0; i < resources.length; ++i)
            output("resource: " + resources[i].har.request.url.replace(/.*((\/[^/]*){3}$)/,"...$1"));
    }
    webInspector.resources.getAll(callbackAndNextTest(onResource, nextTest));
}

function extension_testGetInvalidResource(nextTest)
{
    function onResource(result)
    {
        output("Attempted to retrieve invalid resource: " + JSON.stringify(result));
    }
    webInspector.resources.get(2128506, callbackAndNextTest(onResource, nextTest));
}

function extension_testCreatePanel(nextTest)
{
    function onPanelCreated(panel)
    {
        output("Panel created");
        dumpObject(panel);
    }
    webInspector.panels.create("Test Panel", "extension-panel.html", "extension-panel.png", callbackAndNextTest(onPanelCreated, nextTest));
    output("done createPanel");
}

function extension_testCreateSidebar(nextTest)
{
    function onSidebarCreated(sidebar)
    {
        output("Sidebar created");
        dumpObject(sidebar);
    }
    webInspector.panels.scripts.createSidebarPane("Test Sidebar", "extension-sidebar.html", callbackAndNextTest(onSidebarCreated, nextTest));
}

function extension_testResourceNotification(nextTest)
{
    function onResourceFinished(resource)
    {
        output("Resource finished: " + resource.har.request.url.replace(/.*((\/[^/]*){3}$)/,"...$1"));
    }

    webInspector.resources.onFinished.addListener(callbackAndNextTest(onResourceFinished, nextTest));
    webInspector.inspectedWindow.evaluate("var xhr = new XMLHttpRequest(); xhr.open('GET', '" + location.href + "', false); xhr.send(null);");
}

</script>
</head>
<body onload="runTest()">
<p>Tests WebInspector extension API</p>
</body>
</html>
