LayoutTests/inspector/extensions-expected.txt

Tests WebInspector extension API

Page reloaded.
Started extension.
Running tests...
RUNNING TEST: extension_testCreatePanel
done createPanel
Panel created
{
    createSidebarPane : <function>
    onSelectionChanged : {
        addListener : <function>
        removeListener : <function>
    }
    onSearch : {
        addListener : <function>
        removeListener : <function>
    }
}
RUNNING TEST: extension_testCreateSidebar
Sidebar created
{
    setHeight : <function>
    setExpanded : <function>
}
RUNNING TEST: extension_testGetAllResources
resource: .../tests/inspector/inspector-test2.js
resource: .../LayoutTests/inspector/extensions-test.js
resource: .../LayoutTests/inspector/extensions.html
RUNNING TEST: extension_testGetInvalidResource
Attempted to retrieve invalid resource: {"code":"E_NOTFOUND","description":"Object not found: %s","details":[2128506],"isError":true}
RUNNING TEST: extension_testResourceNotification
Resource finished: .../inspector/resources/extension-main.html
All tests done.
